project_support/Pepsico/requirements.txt
jupyter
mysql-connector-python==8.0.33
# SQLAlchemy==2.0.30
pandas==2.2.2
# pymysql==1.1.0